SPG2XX - Make audio emulation a subdevice
(I need to reuse some of it for SPG110)
